+++ listeners/src/main/java/org/fao/geonet/listener/metadata/draft/ApprovePublishedRecord.java
@@ -23,6 +23,7 @@
 
 package org.fao.geonet.listener.metadata.draft;
 
+import jeeves.server.dispatchers.ServiceManager;
 import org.fao.geonet.constants.Geonet;
 import org.fao.geonet.domain.AbstractMetadata;
 import org.fao.geonet.domain.ISODate;
@@ -69,6 +70,9 @@ public class ApprovePublishedRecord implements ApplicationListener<MetadataPubli
     @Autowired
     IMetadataUtils metadataUtils;
 
+    @Autowired
+    ServiceManager serviceManager;
+
     @Override
     @Transactional
     public void onApplicationEvent(MetadataPublished event) {
@@ -85,6 +89,7 @@ public class ApprovePublishedRecord implements ApplicationListener<MetadataPubli
             // Only do something if the workflow is enabled
             MetadataStatus previousStatus = metadataStatus.getStatus(event.getMd().getId());
             if (previousStatus != null) {
+              try (ServiceContext context = serviceManager.createServiceContext("approve_publish", -1)) {
                 AbstractMetadata publishedMd;
 
                 // On republication the metadata received in the event is the draft version
@@ -99,6 +104,7 @@ public class ApprovePublishedRecord implements ApplicationListener<MetadataPubli
                 }
 
                 draftUtilities.replaceMetadataWithDraft(publishedMd);
+              }
             }
         } catch (Exception e) {
             Log.error(Geonet.DATA_MANAGER, "Error upgrading workflow of " + event.getMd(), e);
