+++ listeners/src/main/java/org/fao/geonet/listener/metadata/draft/DraftRemoved.java
@@ -29,6 +29,8 @@ import java.util.List;
 import javax.transaction.Transactional;
 import javax.transaction.Transactional.TxType;
 
+import jeeves.server.context.ServiceContext;
+import jeeves.server.dispatchers.ServiceManager;
 import org.fao.geonet.constants.Geonet;
 import org.fao.geonet.domain.AbstractMetadata;
 import org.fao.geonet.domain.MetadataDraft;
@@ -58,11 +60,14 @@ public class DraftRemoved {
     @Autowired
     private IMetadataIndexer metadataIndexer;
 
+    @Autowired
+    ServiceManager serviceManager;
+
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMPLETION, fallbackExecution = true)
     public void doAfterCommit(MetadataDraftRemove event) {
         Log.trace(Geonet.DATA_MANAGER, "Reindexing non drafted versions of uuid " + event.getMd().getUuid());
 
-        try {
+        try (ServiceContext context = serviceManager.createServiceContext("draft_removed", -1)) {
             for (AbstractMetadata md : getRecords(event)) {
                 if (!(md instanceof MetadataDraft)) {
                     Log.trace(Geonet.DATA_MANAGER, "Reindexing " + md.getId());
